fix(spa-discovery): isolate wallet cache helpers
